aula6-7/consumer/requirements.txt
redis
mysql-connector-python